필요없는 파일 구분
