--- src/App2.js
@@ -1,12 +1,0 @@
-import { CommonLayOut, GlobalStyle } from "./styles/GlobalStyle";
-
-function App() {
-  return (
-    <>
-      <GlobalStyle />
-      <CommonLayOut></CommonLayOut>
-    </>
-  );
-}
-
-export default App;
